Add light model and diffuse shading
Adds a data structure for directional lights, and makes
modifications to the transformation code to perserve the
positions of the vertices in camera space so they may
be used in shading.

Adds diffuse (Lambertian) shading. Presently this is only
in the red channel, and requires large intensity values
since there is no scaling to the range of pixel values.