Improve user control over model output
This commit adds a number of options for the user to configure the
output and transform the rendered model. In particular, support for
clockwise winding angles has been added, increasing the number of models
that can be rendered.

Models can now be rotated and translated arbitrarily, and scaled
uniformly by an arbitrary amount, using command line options. This is
useful for rendering models that are not centred on the origin or
aligned with the axes.

The lights are now transformed separately to the model, remaining fixed
to the camera while the model rotates, rather than remaining fixed to
the model. A future addition could be to allow transformation of the
lights as well as the model.

The readme is updated to document the new options, and some sample
output is added to the repository to display in the readme.