Add Material class and pass UVs to drawing func
Added a representation of a material, consisting of a diffuse colour and
optionally a diffuse texture supporting linear interpolation based on UV
coordinates. This will allow different faces to have different colours,
and to draw textures across the surface of the faces.

Additionally, update the arguments of the functions for loading a .obj
file and adding a square to have access to materials.