Add texture support and libpng requirement
This completes the work on interpolating texture coordinates and
displaying textures on surfaces, and due to problems encountered on
Windows with ImageMagick, changes CImg to use libpng internally instead
of deferring this to ImageMagick.

Unfortunately, difficulties linking libpng have made this difficult to
fully test on Windows, and it is not certain whether this commit works
properly.