Remove unnecessary use of libpng
Previously, it appeared that on Windows, a problem with ImageMagick or
GraphicsMagick was interfering with the use of PNG files, and an attempt
was made to replace this with CImg's native linking of libpng. On
further investigation, it became apparent first that it was almost
impossible to get libpng to properly link on Windows whether by Conan or
any other means, and that the problem with ImageMagick or GraphicsMagick
was actually a non-fatal warning message. Accordingly, libgpng has been
removed, reverting to use of ImageMagick or GraphicsMagick.