Rewrite drawing functions to be more OpenGL-like
This is an extensive rewrite of most of the functions relating to
drawing.

OpenGL stores the inverse of the camera-space depth value in the w
coordinate of gl_fragcoord, and uses this for perspective-correct
interpolation rather than attempting to use the same depth coordinate as
is used for depth buffer calculations. As well as being more
pedagogically useful, this change should make the perspective-correct
interpolation code simpler and more understandable.

To enable this, in a number of places vec4 is now used instead of
vec3. Some previous instances of std::array<float,3> have been turned
back into glm::vec3 in order to make the operations on them clearer and
more simply stated than is possible in STL.

Flat shading based purely on vertex positions has been temporarily
disabled. This will be re-added by means of a function for calculating
new vertex normals from the face normals, since as far as I know, OpenGL
does not have per-face data such as normals available during
rasterisation and the normals passed to the fragment shader are always
interpolated from vertex normals.