Fix segfault when UV coordinates are missing
While implementing UV coordinate interpolation, I forgot to check for
the case where there are no UV coordinates, in which case a face's index
into the list of UV coordinate pairs is set to be -1 - which inevitably
goes outside the bounds of the vector and causes a segfault. There is
now a check that the UV index is not -1 before attempting to de-index.

This commit also makes a small simplification to the drawing function's
treatment of depths and inverse depths, requiring one less division per
fragment.