Load materials & textures from arbitrary directory
Previously, the program could only load Wavefront .mtl files from its
working directory, which is inconvenient if we want to load .obj files
from another directory and they have their .mtl files with them as they
usually do.

Ideally, we would go automatically from the directory of the .obj file,
but this is surprisingly complicated to achieve, so as a workaround a
command line flag has been added which will allow the user to tell the
program where to look for materials.